Support for regular expressions
By adding in a parser for regular expression syntax as well as
transforms that convert the parsed regular expression output into
instaparse combinators, we can use instagenerate on grammars that
contain regular expressions. This commit adds in some simple support for
the basics of regular experssions, as well as adding some tests.